requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.0.3
